// src/services/generated/monitoring/models/createEmployeeDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Dexa Group Monitoring Employee API
 * REST API admin HRD untuk master karyawan, absensi read-only, audit, dan notifikasi.
 * OpenAPI spec version: 1.0.0
 */

export interface CreateEmployeeDto {
  email: string;
  name: string;
  position: string;
  phoneNumber?: string;
  photoUrl?: string;
  /**
   * Dikirim sekali ke attendance service dan tidak disimpan di monitoring DB
   * @minLength 8
   */
  initialPassword: string;
}